Refactor logic to stop repeated logic.
